requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
google-genai
python-dotenv
langgraph
langsmith
langchain[google-genai]
langchain_community
psycopg[binary]
sentence-transformers
